+++ cloud_functions/vtx-subscriptions/handler.py
@@ -1,22 +1,10 @@
 import json
 import base64
-from twilio.rest import Client
 from urllib import parse
 import boto3
 import time
+from lib import send_sms
 
-ENV = 'dev'
-
-def _load_twilio_client():
-    secrets_client = boto3.session.Session().client(service_name='secretsmanager', region_name="us-east-1")
-    get_secret_value_response = secrets_client.get_secret_value(SecretId='vaccinate-texas')
-    secret = json.loads(get_secret_value_response['SecretString'])
-    return Client(secret['TWILIO_SID'], secret['TWILIO_SECRET'])
-
-twilio_client = _load_twilio_client()
-from_ = "+1 512 488 6383"
-
-# get the message
 def _parse_inbound_msg(event):
     base64_message = event['body']
     base64_bytes = base64_message.encode('ascii')
@@ -24,26 +12,17 @@ def _parse_inbound_msg(event):
     message = message_bytes.decode('ascii')
     return parse.parse_qs(message)
     
-# if body has a valid zip...save and respond confirm
-
-def send_sms(to_, body):
-    if ENV == 'dev':
-        print(f"Message sent {from_} to {to_} -- ENV=DEV:\n", body)
-        return 
-
-    msg = twilio_client.messages.create(body=body,from_=from_,to=to_)
-    resp = f"Message sent {from_} to {to_} with locations -- {msg.sid}"
-    print(resp)
-
 def _extract_valid_zip(body):
     s3_bucket = boto3.resource('s3').Bucket('vtx-public')
     data = json.load(s3_bucket.Object(key='tx_zip_geo_cache.json').get()['Body'])
     valid_zips = data.keys()
     tokens = body.split(" ")
+    print("tokens:\n", tokens)
     for token in tokens:
         if token in valid_zips:
+            print(token)
             return token
-    raise Exception
+    return None
 
 def subscribe(to_, zip_, msgContent, radius = 100):
     data = dict(
@@ -54,42 +33,26 @@ def subscribe(to_, zip_, msgContent, radius = 100):
         meta = msgContent
     )
     s3_bucket = boto3.resource('s3').Bucket('vtx-subscriptions')
-    s3_bucket.Object(key=f"{to_}#{zip_}").put(Body=json.dumps(data))
-    return 'saved'
-
-def send_invalid_repsonse(from_):
-    to_ = from_
-    body = f"Nah that didn't work"
-    send_sms(to_, body)
-    print(body)
-
-def send_subscribe_confirmation(from_, zip_, radius = 100):
-    to_ = from_
-    body = f"Thanks. You're registered to received notifications for {zip_} + {radius} miles."
-    send_sms(to_, body)
-
-def try_suscribe(msgContent):
-    body = msgContent['Body'][0]
-    from_ = body = msgContent['From'][0]
-    
-    zip_ = _extract_valid_zip(body)
-    subscribe(from_, zip_, msgContent)
-    send_subscribe_confirmation(from_, zip_)
-    # except:
-    #     send_invalid_repsonse(from_)
+    subscription_key = f"{to_}#{zip_}"
+    s3_bucket.Object(key=subscription_key).put(Body=json.dumps(data))
+    print("subscription added:", subscription_key)
+    response_sms_body = f"Congrats! You're registered to received notifications for {zip_} + {radius} miles. You're almost ready to #goandgetit"
+    send_sms(to_, response_sms_body)
+    return f"subscription_key {subscription_key} added"
 
 def main(event, context):
-    print("event:", event)
+    # print("event:", event)
     # print("context:", context)
-    
     msgContent = _parse_inbound_msg(event)
-    print("msgContent", msgContent)
-    try_suscribe(msgContent)
-
-    body = {
-        "message": "Go Serverless v1.0! Your function executed successfully!",
-        "input": event
-    }
+    # print("msgContent", msgContent)
+    body = msgContent['Body'][0]
+    from_ = msgContent['From'][0]
+    zip_ = _extract_valid_zip(body)
+    if zip_:
+        subscribe(from_, zip_, msgContent)
+    else: 
+        response_sms_body = f"Hmmm. Doesn't look like you gave us a valid 5 digit zip code."
+        send_sms(from_, response_sms_body)
 
     response = {
         "statusCode": 200,
